Add Python3 support
This (not as large as it looks) commit refactors the codebase to be
compatible with python3, which means moving several files around, and
some small functional changes related to xrange and how binary files
were imported.

Also: A small bug was fixed which creates a default address
group (starting at 0x00) when one is not present.  This allows for intel
hex files created by avr-gcc to be properly parsed (a test was also
added)